feat: Add styling to pie chart
