+++ src/Components/Homepage/Homepage.scss
@@ -14,23 +14,23 @@
   }
 }
 
-.input-message-wrapper {
-  display: flex;
-  height: 100%;
-  width: 50%;
-  justify-content: flex-end;
-  align-items: center;
-}
+// .input-message-wrapper {
+//   display: flex;
+//   height: 100%;
+//   width: 50%;
+//   justify-content: flex-end;
+//   align-items: center;
+// }
 
-.input-wrapper {
-  height: 100%;
-  margin-left: 1rem;
-  input {
-    margin: 0rem;
-    height: 90%;
-    align-self: right;
-  }
-}
+// .input-wrapper {
+//   height: 100%;
+//   margin-left: 1rem;
+//   input {
+//     margin: 0rem;
+//     height: 90%;
+//     align-self: right;
+//   }
+// }
 
 #home-main {
   display: flex;
@@ -53,4 +53,5 @@
 .content-wrapper {
   display: flex;
   width: 90%;
+  height: 85%;
 }
