+++ src/Components/UserInfoContainer/UserInfoContainer.scss
@@ -1,8 +1,38 @@
 @import '../../index.scss';
 
 .user-info-wrapper {
+  border-left: 1px solid $brown;
   width: 30%;
   padding: 1rem;
   @include sourceSans;
-  font-size: 1.5rem;
+  font-size: 1.25rem;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: center;
+  height: 100%;
+  p {
+    text-align: center;
+  }
+  span {
+    @include sourceSansBold;
+  }
+}
+
+.input-wrapper {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+
+  input {
+    margin: 0rem;
+    height: 2rem;
+  }
+  p {
+    font-size: 1rem;
+  }
+}
+
+.pie-chart {
+  height: 50%;
 }
